[refactor] Remove import taichi in taichi/lang/impl.py
ghstack-source-id: d51cdda36f8c8ea1f1b363ab40b48a5dc3871362
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/3894
