+++ python/taichi/__init__.py
@@ -2,7 +2,6 @@ from taichi._funcs import *
 from taichi._logging import *
 from taichi._misc import *
 from taichi.lang import *  # pylint: disable=W0622 # TODO(archibate): It's `taichi.lang.core` overriding `taichi.core`
-from taichi.lang.ndrange import GroupedNDRange, ndrange
 from taichi.main import main
 from taichi.tools import *
 from taichi.tools.patterns import taichi_logo
